% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CI_BCa_Bootstrap.R
\name{CI_BCa_Bootstrap}
\alias{CI_BCa_Bootstrap}
\title{Estimating BCa confidence interval with Bootstrap using R}
\usage{
CI_BCa_Bootstrap(x, conf)
}
\arguments{
\item{x}{the sample}

\item{conf}{confidence level}
}
\value{
a list including the sample estimate and BCa confidence interval
}
\description{
Estimating BCa confidence interval with Bootstrap using R
}
\examples{
\dontrun{
library("bootstrap")
data(scor,package="bootstrap")
CI_BCa_Bootstrap(scor,conf=0.95)
}
}
