% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Metropolis}
\alias{Metropolis}
\title{A random walk Metropolis sampler}
\usage{
Metropolis(sigma, x0, N)
}
\arguments{
\item{x0:}{the initial point(double)}

\item{sigma:}{the standard deviation in the normal distribution(double)}

\item{N:}{the length of the chain(int)}
}
\value{
a random sample of size N
}
\description{
A random walk Metropolis sampler for generating the standard Laplace distribution
}
\examples{
\dontrun{
N<-100
sigma<-1
x0<-0
Metropolis(sigma,x0,N)
}
}
