\name{SC19083-package}
\alias{SC19083-package}
\docType{package}
\title{
  Final homework for 'Statistical Computing' course
}
\description{
  The SC19083 package contains three functions and all homeworks. The three functions are Equity_after_shock, CI_BCa_Bootstrap and Metropolis. Equity_after_shock is used to calculate the residual equity of banking system after a certain external shock using R. CI_BCa_Bootstrap is used to calculate the BCa confidence interval through bootstrap method using R. Metropolis is A random walk Metropolis sampler for generating the standard Laplace distribution using Rcpp.
}
\author{
Shuyue Jin

Maintainer: Shuyue Jin <jean@mail.ustc.edu.cn>
}
\reference{
 "A Bayesian Methodology for Systemic Risk Assessment in Financial Networks"
 "Distress and default contagion in financial networks"
}
\keyword{ package }

